windowscodecs: Remove useless casts to long.

+++ dlls/windowscodecs/ungif.c
@@ -460,8 +460,7 @@ DGifGetImageDesc(GifFileType * GifFile) {
 
     GifFile->ImageCount++;
 
-    Private->PixelCount = (long)GifFile->Image.Width *
-       (long)GifFile->Image.Height;
+    Private->PixelCount = GifFile->Image.Width * GifFile->Image.Height;
 
     DGifSetupDecompress(GifFile);  /* Reset decompress algorithm parameters. */
